+++ models/Absent.js
@@ -14,19 +14,6 @@ module.exports = { absentModel }
 
 
 
-
-const hod_schema = mongoose.Schema(
-    {
-        "HOD_username": { type: String, required: true },
-        "HOD_Password": { type: String, required: true }
-    }
-)
-let hodModel = mongoose.model("Hod", hod_schema)
-module.exports = { hodModel }
-
-
-
-
 const faculty_schema = mongoose.Schema(
     {
         "faculty_username": { type: String, required: true },
